Clean up the existing functionality and make it specific to the prepublish panel
